mlops-backend-main/requirements.txt
fastapi==0.78.0
uvicorn==0.18.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.40.0
psycopg2-binary
deeplake==3.1.12
minio==7.1.13
mlflow==2.1.1
python-gitlab==3.13.0
pyyaml